+++ example/src/example.js
@@ -4,17 +4,17 @@ var Switch = require('react-flexible-switch');
 
 var App = React.createClass({
 	getInitialState() {
-		return { locked: true, externalActive: true }
+		return { locked: true, externalValue: true }
 	},
 
 	render () {
 		return (
 			<div>
 				<h2> On By default </h2>
-				<Switch active />
+				<Switch value={true} />
 
 				<h2> Off By default </h2>
-				<Switch inactive />
+				<Switch value={false} />
 
 				<h2> Custom Colors </h2>
 				<Switch circleStyles={{ onColor: 'blue', offColor: 'red' }} />
@@ -33,23 +33,22 @@ var App = React.createClass({
 				<Switch labels={{ on: 'On', off: 'Off'}} />
 
 				<h2> Locking the Switch </h2>
-				<button onClick={() => this.setState({ locked: false})}>Unlock Switches</button>
-				<button onClick={() => this.setState({ locked: true})}>Lock Switches</button>
+				<button onClick={() => this.setState({ locked: false }) }>Unlock Switch</button>
+				<button onClick={() => this.setState({ locked: true }) }>Lock Switch</button>
 				<br />
 				<br />
 
 				<Switch locked={this.state.locked} />
 				<br />
-				<Switch active locked={this.state.locked} />
 
 				<h2> External Controls </h2>
-				<button onClick={() => this.setState({ externalActive: !this.state.externalActive})}>Toggle Switches</button>
+				<button onClick={() => this.setState({ externalValue: !this.state.externalValue })}>Toggle Switches</button>
 				<br />
 				<br />
 
-				<Switch locked active={this.state.externalActive} />
+				<Switch locked active={this.state.externalValue} onChange={ (value) => { this.setState({ externalValue: value }) }} />
 				<br />
-				<Switch locked active={!this.state.externalActive} />
+				<Switch locked active={!this.state.externalValue} onChange={ (value) => { this.setState({ externalValue: !value }) }} />
 			</div>
 		);
 	}