+++ src/Switch.js
@@ -14,64 +14,31 @@ class Switch extends React.Component {
 
     this.isTouchDevice = window['ontouchstart'] !== undefined;
 
-    const activeState = this.activeStateFromProps(this.props);
-
-    this.state = { sliding: false, active: activeState };
+    this.state = { sliding: false, value: this.props.value };
   }
 
   componentDidMount() {
-    if (!!this.props.locked) {
-      return;
-    }
-
     this.addListener();
   }
 
   componentWillReceiveProps(nextProps) {
-    const lockedChanged = nextProps.locked !== this.props.locked;
-    if (lockedChanged) {
-      nextProps.locked ? this.removeListener() : this.addListener();
+    if (nextProps.value === undefined) {
+      return;
     }
 
-    if (nextProps.active !== this.props.active) {
-      const newActiveState = this.activeStateFromProps(nextProps);
-
-      if (newActiveState !== this.state.active) {
-        this.state = { active: newActiveState };
-      }
+    if (nextProps.value !== this.state.value) {
+      this.setState({ value: nextProps.value })
     }
   }
 
   componentDidUpdate(prevProps, prevState) {
-    if (this.state.active != prevState.active) {
-      const callback = this.state.active ? this.props.onActive : this.props.onInactive;
-      callback && callback();
+    if (this.state.value != prevState.value) {
+      this.props.onChange(this.state.value);
     }
   }
 
   componentWillUnmount() {
-    if (!!this.props.locked) {
-      return;
-    }
-
     this.removeListener();
-  }
-
-  activeStateFromProps (props) {
-    let activeState = false;
-
-    if (typeof props.active == 'undefined' && typeof props.inactive == 'undefined') {
-      activeState = false;
-    }
-
-    if (typeof props.active != 'undefined' && props.active) {
-      activeState = true;
-    }
-
-    if (typeof props.inactive != 'undefined' && props.inactive) {
-      activeState = false;
-    }
-    return activeState;
   }
 
   addListener() {
@@ -95,13 +62,21 @@ class Switch extends React.Component {
   }
 
   onSlideEnd() {
+    if (this.props.locked) {
+      return;
+    }
+
     if (this.state.sliding) {
-      this.setState({ sliding: false, active: !this.state.active });
+      this.setState({ sliding: false, value: !this.state.value });
       reEnableScroll();
     }
   }
 
   onSlideStart(e) {
+    if (this.props.locked) {
+      return;
+    }
+
     if (e.target == this.refs.circle || e.target == this.refs.switch) {
       this.setState({ sliding: true });
       disableScroll();
@@ -116,8 +91,8 @@ class Switch extends React.Component {
     return classNames(
       'switch',
       { sliding: this.state.sliding },
-      { active: this.state.active },
-      { inactive: !this.state.active }
+      { active: this.state.value },
+      { inactive: !this.state.value }
     );
   }
 
@@ -134,9 +109,9 @@ class Switch extends React.Component {
     const switchStyles = this.switchStyles();
 
     const offset = switchStyles.width - circleStyles.diameter;
-    let translation = this.state.active ? offset : 0;
+    let translation = this.state.value ? offset : 0;
 
-    if (this.state.sliding && this.state.active) {
+    if (this.state.sliding && this.state.value) {
       translation -= (circleStyles.diameter / 4 + switchStyles.padding / 4);
     }
 
@@ -147,7 +122,7 @@ class Switch extends React.Component {
 
   backgroundStyle() {
     const circleStyles = this.circleStylesProps();
-    const backgroundColor = this.state.active ? circleStyles.onColor : circleStyles.offColor;
+    const backgroundColor = this.state.value ? circleStyles.onColor : circleStyles.offColor;
     return { backgroundColor };
   }
 
@@ -182,8 +157,8 @@ class Switch extends React.Component {
       ref="switch"
       onMouseLeave={this.onMouseLeave}>
 
-      <Label active={this.state.active} labels={this.props.labels} ref="label" />
-      <span style={this.circleStyles()} className="circle" ref="circle"></span>
+        <Label active={this.state.value} labels={this.props.labels} ref="label" />
+        <span style={this.circleStyles()} className="circle" ref="circle"></span>
       </span>
     );
   }
@@ -209,15 +184,13 @@ const defaultCircleStyles = {
 };
 
 Switch.propTypes = {
-  active: React.PropTypes.bool,
+  value: React.PropTypes.bool,
 
   circleStyles: React.PropTypes.shape({
     onColor: React.PropTypes.string,
     offColor: React.PropTypes.string,
     diameter: React.PropTypes.number
   }),
-
-  inactive: React.PropTypes.bool,
 
   labels: React.PropTypes.shape({
     on: React.PropTypes.string,
@@ -226,8 +199,7 @@ Switch.propTypes = {
 
   locked: React.PropTypes.bool,
 
-  onActive: React.PropTypes.func,
-  onInactive: React.PropTypes.func,
+  onChange: React.PropTypes.func,
 
   switchStyles: React.PropTypes.shape({
     width: React.PropTypes.number
@@ -235,11 +207,11 @@ Switch.propTypes = {
 };
 
 Switch.defaultProps = {
-  onInactive: function() {},
-  onActive: function() {},
+  onChange: (function() {}),
   circleStyles: defaultCircleStyles,
   switchStyles: defaultSwitchStyles,
-  labels: { on: '', off: '' }
+  labels: { on: '', off: '' },
+  locked: false
 };
 
 export default Switch;